[MarkScan] Extract reusable sections from PollWorkerScreen
