+++ libs/mark-flow-ui/src/components/poll_worker/index.ts
@@ -1,3 +1,4 @@
 export * from './ballot_style_select';
 export * from './elements';
 export * from './sections';
+export * from './update_polls_button';
