+++ libs/mark-flow-ui/src/components/poll_worker/sections.tsx
@@ -1,13 +1,88 @@
 /* istanbul ignore file - @preserve - currently tested via apps. */
 
-import { H4 } from '@votingworks/ui';
+import {
+  Font,
+  H2,
+  H3,
+  H4,
+  P,
+  SignedHashValidationApiClient,
+  SignedHashValidationButton,
+  TestModeCallout,
+} from '@votingworks/ui';
+import React from 'react';
+import {
+  format,
+  getPollsStateName,
+  getPollTransitionsFromState,
+} from '@votingworks/utils';
 import {
   Election,
   getAllPrecinctsAndSplits,
+  PollsState,
   PrecinctSelection,
 } from '@votingworks/types';
 import { BallotStyleSelect, OnBallotStyleSelect } from './ballot_style_select';
-import { VotingSession } from './elements';
+import { ButtonGrid, VotingSession } from './elements';
+import { UpdatePollsButton } from './update_polls_button';
+
+export interface HeaderProps {
+  ballotsPrintedCount: number;
+  liveMode: boolean;
+}
+
+export function SectionHeader(props: HeaderProps): JSX.Element {
+  const { ballotsPrintedCount, liveMode: isLiveMode } = props;
+
+  return (
+    <React.Fragment>
+      <div
+        style={{
+          display: 'flex',
+          justifyContent: 'space-between',
+          alignItems: 'start',
+        }}
+      >
+        <H2 as="h1">Poll Worker Menu</H2>
+        {!isLiveMode && <TestModeCallout />}
+      </div>
+      <P>Remove the poll worker card to leave this screen.</P>
+      <P style={{ fontSize: '1.2em' }}>
+        <Font weight="bold">Ballots Printed:</Font>{' '}
+        {format.count(ballotsPrintedCount)}
+      </P>
+    </React.Fragment>
+  );
+}
+
+export interface SectionPollsStateProps {
+  pollsState: PollsState;
+  updatePollsState: (pollsState: PollsState) => void;
+}
+
+export function SectionPollsState(props: SectionPollsStateProps): JSX.Element {
+  const { pollsState, updatePollsState } = props;
+
+  return (
+    <React.Fragment>
+      <H3>
+        Polls: <Font weight="regular">{getPollsStateName(pollsState)}</Font>
+      </H3>
+      <ButtonGrid>
+        {getPollTransitionsFromState(pollsState).map(
+          (pollsTransition, index) => (
+            <UpdatePollsButton
+              pollsTransition={pollsTransition}
+              updatePollsState={updatePollsState}
+              isPrimaryButton={index === 0}
+              key={`${pollsTransition}-button`}
+            />
+          )
+        )}
+      </ButtonGrid>
+    </React.Fragment>
+  );
+}
 
 export interface SectionSessionStartProps {
   election: Election;
@@ -42,5 +117,22 @@ export function SectionSessionStart(
         onSelect={onChooseBallotStyle}
       />
     </VotingSession>
+  );
+}
+
+export interface SectionSystemProps {
+  apiClient: SignedHashValidationApiClient;
+}
+
+export function SectionSystem(props: SectionSystemProps): JSX.Element {
+  const { apiClient } = props;
+
+  return (
+    <React.Fragment>
+      <H3>System</H3>
+      <ButtonGrid>
+        <SignedHashValidationButton apiClient={apiClient} />
+      </ButtonGrid>
+    </React.Fragment>
   );
 }
